Change in Entity Framework Usage
after reading through Module 7 and the SOLID principles, I  switched up the way we were using Entity Framework, to be more conducive with these principles.
I also changed up how we would be using razor syntax and pages, with tag helpers to match the SOLID principles and make it somewhat easier to read the html and razor code.